+++ tests/test_builtins.py
@@ -19,6 +19,7 @@ from xonsh.built_ins import (
     in_macro_call,
     list_of_list_of_strs_outer_product,
     list_of_strs_or_callables,
+    path_literal,
     pathsearch,
     regexsearch,
     reglob,
@@ -48,11 +49,16 @@ def home_env(xession):
 
 
 @skip_if_on_windows
-def test_repath_backslash(home_env):
-    exp = os.listdir(HOME_PATH)
+def test_repath_backslash(home_env, tmp_path):
+    path = tmp_path / "test_repath_backslash"
+    path.mkdir(exist_ok=True, parents=True)
+    (path / ".git").touch()
+    (path / "dir").touch()
+    (path / "file").touch()
+    exp = os.listdir(path)
     exp = {p for p in exp if re.match(r"\w\w.*", p)}
-    exp = {os.path.join(HOME_PATH, p) for p in exp}
-    obs = set(pathsearch(regexsearch, r"~/\w\w.*"))
+    exp = {os.path.join(path, p) for p in exp}
+    obs = set(pathsearch(regexsearch, rf"{path}{os.sep}\w\w.*"))
     assert exp == obs
 
 
@@ -419,3 +425,17 @@ def test_enter_macro():
     assert obj.macro_block == "wakka"
     assert obj.macro_globals
     assert obj.macro_locals
+
+
+def test_xonshpathliteral_contextmanager(tmp_path):
+    start_cwd = os.getcwd()
+    p = path_literal(str(tmp_path))
+    try:
+        with p.cd():
+            assert os.getcwd() == str(p)
+        assert os.getcwd() == start_cwd
+    finally:
+        try:
+            os.chdir(start_cwd)
+        except Exception:
+            pass