+++ xonsh/aliases.py
@@ -7,6 +7,7 @@ import operator
 import os
 import pathlib
 import re
+import shlex
 import shutil
 import sys
 import types
@@ -18,6 +19,7 @@ from typing import Literal
 import xonsh.completers._aliases as xca
 import xonsh.history.main as xhm
 import xonsh.xoreutils.which as xxw
+import xonsh.xoreutils.xcontext as xxt
 from xonsh.built_ins import XSH
 from xonsh.cli_utils import Annotated, Arg, ArgParserAlias
 from xonsh.dirstack import _get_cwd, cd, dirs, popd, pushd
@@ -257,8 +259,10 @@ class Aliases(cabc.MutableMapping):
             except Exception as e:
                 print_exception(f"Exception inside alias {key!r}: {e}")
                 return None
-            if not len(val):
-                raise ValueError("return_command alias: zero arguments.")
+            if not isinstance(val, list) or not len(val):
+                raise ValueError(
+                    f"return_command alias {key!r}: wrong return value {val!r}, expected a list."
+                )
 
         if val is None:
             return default
@@ -669,7 +673,13 @@ def source_foreign_fn(
         if not sourcer:
             return (None, "xonsh: error: `sourcer` command is not mentioned.\n", 1)
         # we have filenames to source
-        prevcmd = "".join([f"{sourcer} {f}\n" for f in files_or_code])
+        shell_name = os.path.basename(shell).lower()
+        quote = (
+            functools.partial(argvquote, force=True)
+            if shell_name in {"cmd", "cmd.exe"}
+            else shlex.quote
+        )
+        prevcmd = "".join(f"{sourcer} {quote(f)}\n" for f in files_or_code)
         files = tuple(files_or_code)
     elif not prevcmd:
         prevcmd = " ".join(files_or_code)  # code to run, no files
@@ -994,22 +1004,30 @@ def trace(args, stdin=None, stdout=None, stderr=None, spec=None):
 
 
 def showcmd(args, stdin=None):
-    """usage: showcmd [-h|--help|cmd args]
+    """usage: showcmd [-e|--expand-alias] [-h|--help] cmd
 
     Displays the command and arguments as a list of strings that xonsh would
     run in subprocess mode. This is useful for determining how xonsh evaluates
     your commands and arguments prior to running these commands.
 
     optional arguments:
+      -e, --expand-alias    expand alias
       -h, --help            show this help message and exit
 
     Examples
     --------
-      >>> showcmd echo $USER "can't" hear "the sea"
+      @ showcmd echo $USER "can't" hear "the sea"
       ['echo', 'I', "can't", 'hear', 'the sea']
+
+      @ aliases['ali'] = 'echo 1'
+      @ showcmd -e ali 2
+      ['echo', '1', '2']
+
     """
     if len(args) == 0 or (len(args) == 1 and args[0] in {"-h", "--help"}):
         print(showcmd.__doc__.rstrip().replace("\n    ", "\n"))
+    elif args[0] in {"-e", "--expand-alias"}:
+        sys.displayhook(XSH.aliases.eval_alias(args[1:]))
     else:
         sys.displayhook(args)
 
@@ -1093,21 +1111,40 @@ def make_default_aliases():
         "trace": trace,
         "timeit": timeit_alias,
         "xonfig": xonfig,
-        "scp-resume": ["rsync", "--partial", "-h", "--progress", "--rsh=ssh"],
         "showcmd": showcmd,
-        "ipynb": ["jupyter", "notebook", "--no-browser"],
         "which": xxw.which,
+        "xcontext": xxt.xcontext,
         "xontrib": xontribs_main,
         "completer": xca.completer_alias,
         "xpip": detect_xpip_alias(),
-        "xonsh-reset": xonsh_reset,
+        "xpython": [XSH.env.get("_", sys.executable)]
+        if IN_APPIMAGE
+        else [sys.executable],
+        "xreset": xonsh_reset,
+        # Command decorators
         "@thread": SpecAttrDecoratorAlias(
             {"threadable": True, "force_threadable": True},
-            "Mark current command as threadable.",
+            "Command decorator. Mark current command as threadable.",
         ),
         "@unthread": SpecAttrDecoratorAlias(
             {"threadable": False, "force_threadable": False},
-            "Mark current command as unthreadable.",
+            "Command decorator. Mark current command as unthreadable.",
+        ),
+        "@lines": SpecAttrDecoratorAlias(
+            {"output_format": "list_lines"},
+            "Command decorator. Return output as list of lines.",
+        ),
+        "@json": SpecAttrDecoratorAlias(
+            {"output_format": lambda lines: XSH.imp.json.loads("\n".join(lines))},
+            "Command decorator. Parses JSON and returns JSON object.",
+        ),
+        "@jsonl": SpecAttrDecoratorAlias(
+            {"output_format": lambda lines: [XSH.imp.json.loads(lj) for lj in lines]},
+            "Command decorator. Parses JSON strings and returns list of JSON objects.",
+        ),
+        "@yaml": SpecAttrDecoratorAlias(
+            {"output_format": lambda lines: XSH.imp.yaml.safe_load("\n".join(lines))},
+            "Command decorator. Parses YAML and returns dict.",
         ),
     }
     if ON_WINDOWS: