+++ xonsh/procs/__init__.py
@@ -1,1 +1,2 @@
+# Fix https://github.com/xonsh/xonsh/pull/5437
 from xonsh.procs import proxies  # noqa